♻️ SWRConfigでfetcherを共通化
