+++ src/pages/_app.tsx
@@ -3,13 +3,25 @@ import type { AppProps } from "next/app";
 import { RecoilRoot } from "recoil";
 import { ThemeProvider } from "@mui/material";
 import { theme } from "../theme";
+import { SWRConfig } from "swr";
+
+const fetcher = async (url: string) => {
+  const response = await fetch(url);
+  if (!response.ok) {
+    throw new Error("データの取得ができなかったので、表示ができません");
+  }
+  const json = await response.json();
+  return json;
+};
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <RecoilRoot>
-      <ThemeProvider theme={theme}>
-        <Component {...pageProps} />
-      </ThemeProvider>
+      <SWRConfig value={{ fetcher }}>
+        <ThemeProvider theme={theme}>
+          <Component {...pageProps} />
+        </ThemeProvider>
+      </SWRConfig>
     </RecoilRoot>
   );
 }
